+++ pysofe/meshes/generation/__init__.py
@@ -1,0 +1,6 @@
+"""
+Provides capabilities for the generation of computational meshes.
+"""
+
+import distance_functions
+import generators